refactor: 게시글 등록/수정/삭제 api와 조회 api는 분리해서 작성
